// 03-graphQL-with-basic-react/4-issue-finder/src/graphql.tsx
export type Maybe<T> = T | null;
export type Exact<T extends { [key: string]: unknown }> = {
  [K in keyof T]: T[K];
};
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  /** An ISO-8601 encoded date string. */
  Date: any;
  /** An ISO-8601 encoded UTC date string. */
  DateTime: any;
  /** A Git object ID. */
  GitObjectID: any;
  /** Git SSH string */
  GitSSHRemote: any;
  /** An ISO-8601 encoded date string. Unlike the DateTime type, GitTimestamp is not converted in UTC. */
  GitTimestamp: any;
  /** A string containing HTML code. */
  HTML: any;
  /** An ISO-8601 encoded UTC date string with millisecond precison. */
  PreciseDateTime: any;
  /** An RFC 3986, RFC 3987, and RFC 6570 (level 4) compliant URI string. */
  URI: any;
  /** A valid x509 certificate string */
  X509Certificate: any;
};
// ...
export type Actor = {
  /** A URL pointing to the actor's public avatar. */
  avatarUrl: Scalars["URI"];
  /** The username of the actor. */
  login: Scalars["String"];
  /** The HTTP path for this actor. */
  resourcePath: Scalars["URI"];
  /** The HTTP URL for this actor. */
  url: Scalars["URI"];
};
// ...
export type Comment = {
  /** The actor who authored the comment. */
  author?: Maybe<Actor>;
  /** Author's association with the subject of the comment. */
  authorAssociation: CommentAuthorAssociation;
  /** The body as Markdown. */
  body: Scalars["String"];
  /** The body rendered to HTML. */
  bodyHTML: Scalars["HTML"];
  /** The body rendered to text. */
  bodyText: Scalars["String"];
  /** Identifies the date and time when the object was created. */
  createdAt: Scalars["DateTime"];
  /** Check if this comment was created via an email reply. */
  createdViaEmail: Scalars["Boolean"];
  /** The actor who edited the comment. */
  editor?: Maybe<Actor>;
  id: Scalars["ID"];
  /** Check if this comment was edited and includes an edit with the creation data */
  includesCreatedEdit: Scalars["Boolean"];
  /** The moment the editor made the last edit */
  lastEditedAt?: Maybe<Scalars["DateTime"]>;
  /** Identifies when the comment was published at. */
  publishedAt?: Maybe<Scalars["DateTime"]>;
  /** Identifies the date and time when the object was last updated. */
  updatedAt: Scalars["DateTime"];
  /** A list of edits to this content. */
  userContentEdits?: Maybe<UserContentEditConnection>;
  /** Did the viewer author this comment. */
  viewerDidAuthor: Scalars["Boolean"];
};

/** Represents a comment. */
export type CommentUserContentEditsArgs = {
  after?: Maybe<Scalars["String"]>;
  before?: Maybe<Scalars["String"]>;
  first?: Maybe<Scalars["Int"]>;
  last?: Maybe<Scalars["Int"]>;
};

/** A comment author association with repository. */
export enum CommentAuthorAssociation {
  /** Author is a member of the organization that owns the repository. */
  Member = "MEMBER",
  /** Author is the owner of the repository. */
  Owner = "OWNER",
  /** Author is a placeholder for an unclaimed user. */
  Mannequin = "MANNEQUIN",
  /** Author has been invited to collaborate on the repository. */
  Collaborator = "COLLABORATOR",
  /** Author has previously committed to the repository. */
  Contributor = "CONTRIBUTOR",
  /** Author has not previously committed to the repository. */
  FirstTimeContributor = "FIRST_TIME_CONTRIBUTOR",
  /** Author has not previously committed to GitHub. */
  FirstTimer = "FIRST_TIMER",
  /** Author has no association with the repository. */
  None = "NONE",
}

/** The possible errors that will prevent a user from updating a comment. */
export enum CommentCannotUpdateReason {
  /** Unable to create comment because repository is archived. */
  Archived = "ARCHIVED",
  /** You must be the author or have write access to this repository to update this comment. */
  InsufficientAccess = "INSUFFICIENT_ACCESS",
  /** Unable to create comment because issue is locked. */
  Locked = "LOCKED",
  /** You must be logged in to update this comment. */
  LoginRequired = "LOGIN_REQUIRED",
  /** Repository is under maintenance. */
  Maintenance = "MAINTENANCE",
  /** At least one email address must be verified to update this comment. */
  VerifiedEmailRequired = "VERIFIED_EMAIL_REQUIRED",
  /** You cannot update this comment */
  Denied = "DENIED",
}
// ...
export type Deletable = {
  /** Check if the current viewer can delete this object. */
  viewerCanDelete: Scalars["Boolean"];
};
// ...
export type IssueComment = Node &
  Comment &
  Deletable &
  Minimizable &
  Updatable &
  UpdatableComment &
  Reactable &
  RepositoryNode & {
    __typename?: "IssueComment";
    /** The actor who authored the comment. */
    author?: Maybe<Actor>;
    /** Author's association with the subject of the comment. */
    authorAssociation: CommentAuthorAssociation;
    /** The body as Markdown. */
    body: Scalars["String"];
    /** The body rendered to HTML. */
    bodyHTML: Scalars["HTML"];
    /** The body rendered to text. */
    bodyText: Scalars["String"];
    /** Identifies the date and time when the object was created. */
    createdAt: Scalars["DateTime"];
    /** Check if this comment was created via an email reply. */
    createdViaEmail: Scalars["Boolean"];
    /** Identifies the primary key from the database. */
    databaseId?: Maybe<Scalars["Int"]>;
    /** The actor who edited the comment. */
    editor?: Maybe<Actor>;
    id: Scalars["ID"];
    /** Check if this comment was edited and includes an edit with the creation data */
    includesCreatedEdit: Scalars["Boolean"];
    /** Returns whether or not a comment has been minimized. */
    isMinimized: Scalars["Boolean"];
    /** Identifies the issue associated with the comment. */
    issue: Issue;
    /** The moment the editor made the last edit */
    lastEditedAt?: Maybe<Scalars["DateTime"]>;
    /** Returns why the comment was minimized. */
    minimizedReason?: Maybe<Scalars["String"]>;
    /** Identifies when the comment was published at. */
    publishedAt?: Maybe<Scalars["DateTime"]>;
    /**
     * Returns the pull request associated with the comment, if this comment was made on a
     * pull request.
     */
    pullRequest?: Maybe<PullRequest>;
    /** A list of reactions grouped by content left on the subject. */
    reactionGroups?: Maybe<Array<ReactionGroup>>;
    /** A list of Reactions left on the Issue. */
    reactions: ReactionConnection;
    /** The repository associated with this node. */
    repository: Repository;
    /** The HTTP path for this issue comment */
    resourcePath: Scalars["URI"];
    /** Identifies the date and time when the object was last updated. */
    updatedAt: Scalars["DateTime"];
    /** The HTTP URL for this issue comment */
    url: Scalars["URI"];
    /** A list of edits to this content. */
    userContentEdits?: Maybe<UserContentEditConnection>;
    /** Check if the current viewer can delete this object. */
    viewerCanDelete: Scalars["Boolean"];
    /** Check if the current viewer can minimize this object. */
    viewerCanMinimize: Scalars["Boolean"];
    /** Can user react to this subject */
    viewerCanReact: Scalars["Boolean"];
    /** Check if the current viewer can update this object. */
    viewerCanUpdate: Scalars["Boolean"];
    /** Reasons why the current viewer can not update this comment. */
    viewerCannotUpdateReasons: Array<CommentCannotUpdateReason>;
    /** Did the viewer author this comment. */
    viewerDidAuthor: Scalars["Boolean"];
  };
// ...
export type Mannequin = Node &
  Actor &
  UniformResourceLocatable & {
    __typename?: "Mannequin";
    /** A URL pointing to the GitHub App's public avatar. */
    avatarUrl: Scalars["URI"];
    /** Identifies the date and time when the object was created. */
    createdAt: Scalars["DateTime"];
    /** Identifies the primary key from the database. */
    databaseId?: Maybe<Scalars["Int"]>;
    /** The mannequin's email on the source instance. */
    email?: Maybe<Scalars["String"]>;
    id: Scalars["ID"];
    /** The username of the actor. */
    login: Scalars["String"];
    /** The HTML path to this resource. */
    resourcePath: Scalars["URI"];
    /** Identifies the date and time when the object was last updated. */
    updatedAt: Scalars["DateTime"];
    /** The URL to this resource. */
    url: Scalars["URI"];
  };
// ...
export type Minimizable = {
  /** Returns whether or not a comment has been minimized. */
  isMinimized: Scalars["Boolean"];
  /** Returns why the comment was minimized. */
  minimizedReason?: Maybe<Scalars["String"]>;
  /** Check if the current viewer can minimize this object. */
  viewerCanMinimize: Scalars["Boolean"];
};
// ...
export type Node = {
  /** ID of the object. */
  id: Scalars["ID"];
};
// ...
export type Reactable = {
  /** Identifies the primary key from the database. */
  databaseId?: Maybe<Scalars["Int"]>;
  id: Scalars["ID"];
  /** A list of reactions grouped by content left on the subject. */
  reactionGroups?: Maybe<Array<ReactionGroup>>;
  /** A list of Reactions left on the Issue. */
  reactions: ReactionConnection;
  /** Can user react to this subject */
  viewerCanReact: Scalars["Boolean"];
};

/** Represents a subject that can be reacted on. */
export type ReactableReactionsArgs = {
  after?: Maybe<Scalars["String"]>;
  before?: Maybe<Scalars["String"]>;
  first?: Maybe<Scalars["Int"]>;
  last?: Maybe<Scalars["Int"]>;
  content?: Maybe<ReactionContent>;
  orderBy?: Maybe<ReactionOrder>;
};

/** The connection type for User. */
export type ReactingUserConnection = {
  __typename?: "ReactingUserConnection";
  /** A list of edges. */
  edges?: Maybe<Array<Maybe<ReactingUserEdge>>>;
  /** A list of nodes. */
  nodes?: Maybe<Array<Maybe<User>>>;
  /** Information to aid in pagination. */
  pageInfo: PageInfo;
  /** Identifies the total count of items in the connection. */
  totalCount: Scalars["Int"];
};
// ...
export type ReactionConnection = {
  __typename?: "ReactionConnection";
  /** A list of edges. */
  edges?: Maybe<Array<Maybe<ReactionEdge>>>;
  /** A list of nodes. */
  nodes?: Maybe<Array<Maybe<Reaction>>>;
  /** Information to aid in pagination. */
  pageInfo: PageInfo;
  /** Identifies the total count of items in the connection. */
  totalCount: Scalars["Int"];
  /** Whether or not the authenticated user has left a reaction on the subject. */
  viewerHasReacted: Scalars["Boolean"];
};

/** Emojis that can be attached to Issues, Pull Requests and Comments. */
export enum ReactionContent {
  /** Represents the `:+1:` emoji. */
  ThumbsUp = "THUMBS_UP",
  /** Represents the `:-1:` emoji. */
  ThumbsDown = "THUMBS_DOWN",
  /** Represents the `:laugh:` emoji. */
  Laugh = "LAUGH",
  /** Represents the `:hooray:` emoji. */
  Hooray = "HOORAY",
  /** Represents the `:confused:` emoji. */
  Confused = "CONFUSED",
  /** Represents the `:heart:` emoji. */
  Heart = "HEART",
  /** Represents the `:rocket:` emoji. */
  Rocket = "ROCKET",
  /** Represents the `:eyes:` emoji. */
  Eyes = "EYES",
}

/** An edge in a connection. */
export type ReactionEdge = {
  __typename?: "ReactionEdge";
  /** A cursor for use in pagination. */
  cursor: Scalars["String"];
  /** The item at the end of the edge. */
  node?: Maybe<Reaction>;
};

/** A group of emoji reactions to a particular piece of content. */
export type ReactionGroup = {
  __typename?: "ReactionGroup";
  /** Identifies the emoji reaction. */
  content: ReactionContent;
  /** Identifies when the reaction was created. */
  createdAt?: Maybe<Scalars["DateTime"]>;
  /** The subject that was reacted to. */
  subject: Reactable;
  /** Users who have reacted to the reaction subject with the emotion represented by this reaction group */
  users: ReactingUserConnection;
  /** Whether or not the authenticated user has left a reaction on the subject. */
  viewerHasReacted: Scalars["Boolean"];
};
// ...
export type RepositoryNode = {
  /** The repository associated with this node. */
  repository: Repository;
};
// ...
export type Updatable = {
  /** Check if the current viewer can update this object. */
  viewerCanUpdate: Scalars["Boolean"];
};

/** Comments that can be updated. */
export type UpdatableComment = {
  /** Reasons why the current viewer can not update this comment. */
  viewerCannotUpdateReasons: Array<CommentCannotUpdateReason>;
};
// ...
export type UserContentEditConnection = {
  __typename?: "UserContentEditConnection";
  /** A list of edges. */
  edges?: Maybe<Array<Maybe<UserContentEditEdge>>>;
  /** A list of nodes. */
  nodes?: Maybe<Array<Maybe<UserContentEdit>>>;
  /** Information to aid in pagination. */
  pageInfo: PageInfo;
  /** Identifies the total count of items in the connection. */
  totalCount: Scalars["Int"];
};

/** An edge in a connection. */
export type UserContentEditEdge = {
  __typename?: "UserContentEditEdge";
  /** A cursor for use in pagination. */
  cursor: Scalars["String"];
  /** The item at the end of the edge. */
  node?: Maybe<UserContentEdit>;
};